feat: a cell can be selected by clicking on it
When a cell is clicked, it will be shown as selected by a "glowing" effect and golden border. The colour of the cell also changes to "lime" for now, just to see the difference and simple tests. Google login code is temporarily disabled so it doesn't interfere with this feature as there were errors.